Move `ui::core::widget` to `ui::widget`.
I have also written documentation for the `slider` widget.
