+++ src/ui/core/node.rs
@@ -8,12 +8,12 @@ use crate::ui::core::{Number, Size, Style};
 /// runtime obtains a [`Node`] by calling [`Widget::node`].
 ///
 /// [`Style`]: struct.Style.html
-/// [`Widget`]: widget/trait.Widget.html
+/// [`Widget`]: trait.Widget.html
 /// [`Node`]: struct.Node.html
-/// [`Widget::node`]: widget/trait.Widget.html#tymethod.node
+/// [`Widget::node`]: trait.Widget.html#tymethod.node
 /// [`Layout`]: struct.Layout.html
 #[derive(Debug)]
-pub struct Node(pub(super) node::Node);
+pub struct Node(pub(crate) node::Node);
 
 impl Node {
     /// Creates a new [`Node`] with the given [`Style`].
@@ -47,7 +47,7 @@ impl Node {
     ///
     /// [`Node`]: struct.Node.html
     /// [`Style`]: struct.Style.html
-    /// [`Widget`]: widget/trait.Widget.html
+    /// [`Widget`]: trait.Widget.html
     pub fn with_measure<F>(style: Style, measure: F) -> Node
     where
         F: 'static + Fn(Size<Number>) -> Size<f32>,
