+++ src/ui/core/widget.rs
@@ -1,21 +1,3 @@
-//! Use the built-in widgets or build your own.
-mod column;
-mod row;
-
-pub mod button;
-pub mod checkbox;
-pub mod radio;
-pub mod slider;
-pub mod text;
-
-pub use button::Button;
-pub use checkbox::Checkbox;
-pub use column::Column;
-pub use radio::Radio;
-pub use row::Row;
-pub use slider::Slider;
-pub use text::Text;
-
 use crate::graphics::Point;
 use crate::ui::core::{Event, Hasher, Layout, MouseCursor, Node};
